requirements.txt
pydantic
aiogram
pytz
supabase==2.10.0
gspread
google-auth
python-dotenv